requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
azure-core==1.30.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
executing==2.0.1
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
ipykernel==6.29.2
ipython==8.21.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.12.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
wcwidth==0.2.13
